TRANSITION: git move unchanged c3, rm unused imports
