--- legacy/c3_trajectory_generator/srv/SetDisabled.srv
@@ -1,2 +1,0 @@
-bool disabled
----
