+++ mil_common/mil_c3_tgen/include/AttitudeHelpers.h
@@ -7,18 +7,18 @@ using namespace Eigen;
 
 namespace subjugator {
 class AttitudeHelpers {
-public:
+ public:
   static Vector3d LocalGravity(double lat, double depth);
   static double Markov_wStdDev(double dt, double T, double sigma);
-  static Vector3d RotationToEuler(const Matrix3d &R);
-  static Matrix3d EulerToRotation(const Vector3d &rpy);
+  static Vector3d RotationToEuler(const Matrix3d& R);
+  static Matrix3d EulerToRotation(const Vector3d& rpy);
   static double DAngleDiff(double a, double b);
   static double DAngleClamp(double a);
-  static Vector4d RotationToQuaternion(const Matrix3d &R);
-  static Matrix3d VectorSkew3(const Vector3d &v);
-  static MatrixXd DiagMatrixFromVector(const VectorXd &v);
-  static VectorXd Tanh(const VectorXd &v);
-  static VectorXd Sech(const VectorXd &v);
+  static Vector4d RotationToQuaternion(const Matrix3d& R);
+  static Matrix3d VectorSkew3(const Vector3d& v);
+  static MatrixXd DiagMatrixFromVector(const VectorXd& v);
+  static VectorXd Tanh(const VectorXd& v);
+  static VectorXd Sech(const VectorXd& v);
 };
 }
