+++ mil_common/mil_c3_tgen/src/AttitudeHelpers.cpp
@@ -4,7 +4,7 @@ using namespace subjugator;
 using namespace Eigen;
 using namespace std;
 
-Matrix3d AttitudeHelpers::EulerToRotation(const Vector3d &rpy) {
+Matrix3d AttitudeHelpers::EulerToRotation(const Vector3d& rpy) {
   double sphi = sin(rpy(0));
   double cphi = cos(rpy(0));
 
@@ -15,8 +15,6 @@ Matrix3d AttitudeHelpers::EulerToRotation(const Vector3d &rpy) {
   double cpsi = cos(rpy(2));
 
   return (Matrix3d() << cpsi * ctheta, -spsi * cphi + cpsi * stheta * sphi,
-          spsi * sphi + cphi * cphi * stheta, spsi * ctheta,
-          cpsi * cphi + sphi * stheta * spsi,
-          -cpsi * sphi + stheta * spsi * cphi, -stheta, ctheta * sphi,
-          ctheta * cphi).finished();
+          spsi * sphi + cphi * cphi * stheta, spsi * ctheta, cpsi * cphi + sphi * stheta * spsi,
+          -cpsi * sphi + stheta * spsi * cphi, -stheta, ctheta * sphi, ctheta * cphi).finished();
 }
