+++ mil_common/mil_c3_tgen/src/C3Trajectory.cpp
@@ -9,16 +9,17 @@ using namespace boost;
 using namespace std;
 
 static double sign(double x) {
-  if (x > 0)
-    return 1;
-  if (x < 0)
-    return -1;
+  if (x > 0) return 1;
+  if (x < 0) return -1;
   return 0;
 }
 
 C3Trajectory::C3Trajectory(const Point &start, const Limits &limits)
-    : q(start.q), qdot(start.qdot), qdotdot_b(Vector6d::Zero()),
-      u_b(Vector6d::Zero()), limits(limits) {}
+    : q(start.q),
+      qdot(start.qdot),
+      qdotdot_b(Vector6d::Zero()),
+      u_b(Vector6d::Zero()),
+      limits(limits) {}
 
 static Vector6d apply(const Matrix4d &T, const Vector6d &q, double w) {
   Vector6d q_t;
@@ -29,13 +30,11 @@ static Vector6d apply(const Matrix4d &T, const Vector6d &q, double w) {
 }
 
 C3Trajectory::PointWithAcceleration C3Trajectory::getCurrentPoint() const {
-  return PointWithAcceleration(
-      q, qdot,
-      apply(C3Trajectory::transformation_pair(q).second, qdotdot_b, 0));
+  return PointWithAcceleration(q, qdot,
+                               apply(C3Trajectory::transformation_pair(q).second, qdotdot_b, 0));
 }
 
-void C3Trajectory::update(double dt, const Waypoint &waypoint,
-                          double waypoint_t) {
+void C3Trajectory::update(double dt, const Waypoint &waypoint, double waypoint_t) {
   pair<Matrix4d, Matrix4d> Ts = transformation_pair(q);
   const Matrix4d &T = Ts.first;
   const Matrix4d &T_inv = Ts.second;
@@ -58,8 +57,7 @@ void C3Trajectory::update(double dt, const Waypoint &waypoint,
   }
 
   if (posdelta.norm() > 0.01 && waypoint.coordinate_unaligned) {
-    pair<Vector3d, Vector3d> result =
-        limit(vmin_b_prime.head(3), vmax_b_prime.head(3), posdelta);
+    pair<Vector3d, Vector3d> result = limit(vmin_b_prime.head(3), vmax_b_prime.head(3), posdelta);
     vmin_b_prime.head(3) = result.first;
     vmax_b_prime.head(3) = result.second;
   }
@@ -75,16 +73,13 @@ void C3Trajectory::update(double dt, const Waypoint &waypoint,
   }
 
   for (int i = 3; i < 6; i++) {
-    while (r_b(i) - q_b(i) > M_PI)
-      r_b(i) -= 2 * M_PI;
-    while (r_b(i) - q_b(i) < -M_PI)
-      r_b(i) += 2 * M_PI;
+    while (r_b(i) - q_b(i) > M_PI) r_b(i) -= 2 * M_PI;
+    while (r_b(i) - q_b(i) < -M_PI) r_b(i) += 2 * M_PI;
   }
 
   for (int i = 0; i < 6; i++) {
-    u_b(i) = c3filter(q_b(i), qdot_b(i), qdotdot_b(i), r_b(i), rdot_b(i), 0,
-                      vmin_b_prime(i), vmax_b_prime(i), amin_b_prime(i),
-                      amax_b_prime(i), limits.umax_b(i));
+    u_b(i) = c3filter(q_b(i), qdot_b(i), qdotdot_b(i), r_b(i), rdot_b(i), 0, vmin_b_prime(i),
+                      vmax_b_prime(i), amin_b_prime(i), amax_b_prime(i), limits.umax_b(i));
   }
 
   qdotdot_b += dt * u_b;
@@ -93,10 +88,8 @@ void C3Trajectory::update(double dt, const Waypoint &waypoint,
   q += dt * qdot;
 
   for (int i = 3; i < 6; i++) {
-    while (q(i) > M_PI)
-      q(i) -= 2 * M_PI;
-    while (q(i) < -M_PI)
-      q(i) += 2 * M_PI;
+    while (q(i) > M_PI) q(i) -= 2 * M_PI;
+    while (q(i) < -M_PI) q(i) += 2 * M_PI;
   }
 }
 
@@ -109,19 +102,16 @@ static double ucv(double v, double edot, double edotdot, double umax) {
   return -umax * sign(tmp + (1 - abs(sign(tmp))) * edotdot);
 }
 
-static double ua(double a, double edotdot, double umax) {
-  return -umax * sign(edotdot - a);
-}
+static double ua(double a, double edotdot, double umax) { return -umax * sign(edotdot - a); }
 
-static double uv(double v, double edot, double edotdot, double edotdotmin,
-                 double edotdotmax, double umax) {
+static double uv(double v, double edot, double edotdot, double edotdotmin, double edotdotmax,
+                 double umax) {
   return max(ua(edotdotmin, edotdot, umax),
              min(ucv(v, edot, edotdot, umax), ua(edotdotmax, edotdot, umax)));
 }
 
-double C3Trajectory::c3filter(double q, double qdot, double qdotdot, double r,
-                              double rdot, double rdotdot, double vmin,
-                              double vmax, double amin, double amax,
+double C3Trajectory::c3filter(double q, double qdot, double qdotdot, double r, double rdot,
+                              double rdotdot, double vmin, double vmax, double amin, double amax,
                               double umax) {
   double e = (q - r) / umax;
   double edot = (qdot - rdot) / umax;
@@ -151,8 +141,7 @@ double C3Trajectory::c3filter(double q, double qdot, double qdotdot, double r,
         sd / 4 * sqrt(2 * pow(pow(edotdot, 2) + 2 * edot * sd, 3));
   }
 
-  double uc =
-      -umax * sign(S + (1 - abs(sign(S))) * (delta + (1 - abs(sd)) * edotdot));
+  double uc = -umax * sign(S + (1 - abs(sign(S))) * (delta + (1 - abs(sd)) * edotdot));
   double uv_emin = uv(edotmin, edot, edotdot, edotdotmin, edotdotmax, umax);
   double uv_emax = uv(edotmax, edot, edotdot, edotdotmin, edotdotmax, umax);
 
@@ -170,16 +159,15 @@ pair<Matrix4d, Matrix4d> C3Trajectory::transformation_pair(const Vector6d &q) {
   T.block<3, 1>(0, 3) = -q.head(3);
 
   pair<Matrix4d, Matrix4d> result;
-  result.first = R.transpose() * T; // NED -> BODY
+  result.first = R.transpose() * T;  // NED -> BODY
 
   T.block<3, 1>(0, 3) = q.head(3);
-  result.second = T * R; // BODY -> NED
+  result.second = T * R;  // BODY -> NED
 
   return result;
 }
 
-std::pair<Vector3d, Vector3d> C3Trajectory::limit(const Vector3d &vmin,
-                                                  const Vector3d &vmax,
+std::pair<Vector3d, Vector3d> C3Trajectory::limit(const Vector3d &vmin, const Vector3d &vmax,
                                                   const Vector3d &delta) {
   Vector3d adelta = delta.array().abs();
